requirements.txt
keras
flask
flask-cors
pillow